% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center}
\alias{center}
\title{Create a column of centered variables}
\usage{
center(x, variables = c(colnames(x)), standardized = FALSE)
}
\arguments{
\item{x}{dataframe}

\item{variables}{c() of columns to standardize}

\item{standardized}{Logical. Do you want to calculate zscores? (Default = FALSE)}
}
\description{
This function centers a variable around the mean. To create z-scores, specify `standardized=TRUE`
A new column will be created with the centered values.
There is also the option to center within context.
}
\examples{
center(x, subset = c(colnames(x)), standardized = FALSE)
}
\keyword{center}
