% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.categorical.R
\name{code.categorical-deprecated}
\alias{code.categorical-deprecated}
\title{Recode a character variable into numerical values}
\usage{
code.categorical(x, variable = "", order = c())
}
\arguments{
\item{x}{dataframe}

\item{variable}{column name}

\item{order}{c() of column values in order}
}
\description{
This function will code a column containing categorical "string" values into numerical values
}
\seealso{
\code{\link{datawrangling-deprecated}}
}
\keyword{internal}
