% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.categorical.R
\name{code.categorical}
\alias{code.categorical}
\title{A Data Transformation Function}
\usage{
code.categorical(x, variable = "", order = c())
}
\arguments{
\item{x}{dataframe}

\item{variable}{column name}

\item{order}{c() of column values}
}
\description{
This function will code a column containing categorical "string" values into numerical values
}
\examples{
code.categorical(x, variable = "columnName", order = c("string1", "string2", "string3"))
}
\keyword{code}
