% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_categorical.R
\name{code_categorical}
\alias{code_categorical}
\title{Recode a character variable into numerical values}
\usage{
code_categorical(x, variable = "", order = c())
}
\arguments{
\item{x}{dataframe}

\item{variable}{column name}

\item{order}{c() of column values in order}
}
\description{
This function will code a column containing categorical "string" values into numerical values
}
\examples{
code_categorical(x, variable = "columnName", order = c("string1", "string2", "string3"))
}
\keyword{code}
