% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite.R
\name{composite}
\alias{composite}
\title{A Data Transformation Function}
\usage{
composite(x, variables, comp = "mean", standardized = FALSE,
  name = NULL, missing.allowed = NULL)
}
\arguments{
\item{x}{dataframe}

\item{variables}{c() of columns to average together to create the composite}

\item{comp}{How the composite should be calculated, i.e. mean or sum. (Default = "mean").}

\item{name}{Name of the new composite variable}

\item{missing.allowed}{Criteria for how many variables can having missing values to not add to composite variable}

\item{standardize}{Logical. Do you want to calculate the composite based on standardized (z-score) values? (Default = FALSE)}
}
\description{
This function will create a composite variable
}
\examples{
composite(x, variable = c(), comp = "mean", name = "name")
}
\keyword{composite}
