% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_study.R
\name{create_study}
\alias{create_study}
\title{Create Directory Organization For a New Study}
\usage{
create_study(scripts.dir = TRUE)
}
\arguments{
\item{scripts.dir}{Logical. Create script directory? default = TRUE}
}
\description{
This function can be used to automatically setup your study directory by
creating folders and a template masterscript
}
\examples{

}
\keyword{center}
