% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates_remove.R
\name{duplicates_remove}
\alias{duplicates_remove}
\title{Remove duplicate subjects from an edat file}
\usage{
duplicates_remove(x, timing.variables = "SessionStartDateTimeUtc",
  subj.variable = "Subject", taskname = "", output.folder = "")
}
\arguments{
\item{x}{Imported E-Merge txt file}

\item{timing.variables}{Variable name in file that corresponds to the session date and time}

\item{subj.variable}{Name of varaible with Subject information}

\item{taskname}{name of task for saving output file}

\item{output.folder}{Folder path to output the duplicate file to}
}
\description{
This function removes duplicate subject files from an E-Merge txt file
}
\examples{
duplicates_remove(x, subset = c(colnames(x)), context = "Subject")
}
