% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.bind.R
\name{files.bind-deprecated}
\alias{files.bind-deprecated}
\title{Bind multiple files into one file}
\usage{
files.bind(path = "", pattern = "", delim = "\\t", na = "",
  output.file = "", bind = "rows")
}
\arguments{
\item{path}{Folder location of files to be merged}

\item{pattern}{Pattern to identify files to be merged}

\item{delim}{Delimiter used in files. Passed onto `readr::read_delim()`}

\item{na}{How are missing values defined in files to be merged. Passed to `readr::write_delim()`}

\item{output.file}{File name and path to be saved to.}
}
\description{
This function binds multiple files together. `bind` can occur by "rows" or "columns"
}
\seealso{
\code{\link{datawrangling-deprecated}}
}
\keyword{internal}
