% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.bindr.R
\name{files.bindr}
\alias{files.bindr}
\title{A Data Transformation Function}
\usage{
files.bindr(path = "", pattern = "", delim = "\\t", na = "",
  output.file = "", bind = "rows")
}
\arguments{
\item{path}{Folder location of files to be merged}

\item{pattern}{Pattern to identify files to be merged}

\item{delim}{Delimiter used in files to be merged}

\item{na}{How are missing values defined in files to be merged}

\item{output.file}{File name and path to be saved}
}
\description{
This function merges files of the same data structure
}
\examples{
files.bindr(path = "./Data", pattern = ".txt")
}
