% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.join.R
\name{files.join-deprecated}
\alias{files.join-deprecated}
\title{Merge multiple files into one merged file}
\usage{
files.join(path = "", pattern = "", delim = "\\t", na = "",
  output.file = "", id = NULL)
}
\arguments{
\item{path}{Folder location of files to be merged}

\item{pattern}{Pattern to identify files to be merged}

\item{delim}{Delimiter used in files. Passed onto `readr::read_delim()`}

\item{na}{How are missing values defined in files to be merged. Passed to `readr::write_delim()`}

\item{output.file}{File name and path to be saved to}

\item{id}{Subject ID variable name. Passed onto `plyr::join_all(by = id)`}
}
\description{
This function merges multiple files together.
}
\seealso{
\code{\link{datawrangling-deprecated}}
}
\keyword{internal}
