% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.merge.R
\name{files.merge}
\alias{files.merge}
\title{A Data Transformation Function}
\usage{
files.merge(path = "", pattern = "", delim = "\\t", na = "NA",
  output.folder = "")
}
\arguments{
\item{path}{Folder location of files to be merged}

\item{pattern}{Pattern to identify files to be merged}

\item{delim}{Delimiter used in files to be merged}

\item{na}{How are missing values defined in files to be merged}
}
\description{
This function merges files of the same data structure
}
\examples{
files.merge(path = "./Data", pattern = ".txt")
}
