% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_bind.R
\name{files_bind}
\alias{files_bind}
\title{Bind multiple files into one file}
\usage{
files_bind(
  path = "",
  pattern = "",
  delim = ",",
  output.delim = ",",
  output.file = "",
  bind = "rows",
  guess_max = 1000
)
}
\arguments{
\item{path}{Folder location of files to be merged}

\item{pattern}{Pattern to identify files to be merged}

\item{delim}{Delimiter used in files. Passed onto `readr::read_delim()`}

\item{output.file}{File name and path to be saved to.}

\item{guess_max}{Maximum}
}
\description{
This function binds multiple files together. `bind` can occur by "rows" or "columns"
}
\examples{
files_bind(path = "./Data", pattern = ".txt")
}
