% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_join.R
\name{files_join}
\alias{files_join}
\title{Merge multiple files into one merged file}
\usage{
files_join(
  path = "",
  pattern = "",
  delim = ",",
  output.delim = ",",
  na = "",
  output.file = "",
  id = NULL
)
}
\arguments{
\item{path}{Folder location of files to be merged}

\item{pattern}{Pattern to identify files to be merged}

\item{delim}{Delimiter used in files. Passed onto `readr::read_delim()`}

\item{output.delim}{Delmiter to use when saving data to file (default = ",")}

\item{na}{How are missing values defined in files to be merged. Passed to `readr::write_delim()`}

\item{output.file}{File name and path to be saved to}

\item{id}{Subject ID variable name. Passed onto `plyr::join_all(by = id)`}
}
\description{
This function merges multiple files together.
}
\examples{
files_join(path = "./Data", pattern = ".txt")
}
