% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_latent.R
\name{remove_latent}
\alias{remove_latent}
\title{Remove subjects with too much missing data on a given latent construct}
\usage{
remove_latent(
  x,
  factor.list,
  missing.allowed,
  id,
  output.dir = NULL,
  output.file = NULL
)
}
\arguments{
\item{x}{dataframe}

\item{factor.list}{list of factors and tasks. ex. `list(WMC = c("OSpan", "RotSpan", SymSpan"), Gf = c("RAPM", "NumberSeries", "LetterSets"))`}

\item{missing.allowed}{Proportion of tasks allowed to be missing}

\item{id}{Subject ID variable}

\item{output.dir}{File directory to save removed subjects to}

\item{output.file}{File name to save removed subjects to}
}
\description{
This function will remove subjects based on amount of missing data per latent construct
}
\examples{
# Example
}
\keyword{remove}
