% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.long.R, R/spread.long.R
\name{reshape.long}
\alias{reshape.long}
\title{Data Transformation Function}
\usage{
reshape.long(x, variables = "", variables_combine.sep = "_",
  values = c(), id = "", fill = NA)

reshape.long(x, variables = "", variables_combine.sep = "_",
  values = c(), id = "", fill = NA)
}
\arguments{
\item{x}{dataframe}

\item{variables}{The variable used for spreading}

\item{variables_combine.sep}{if using more than one variables column then specifiy how values should be seperated when combined}

\item{values}{A vector of columns that contain the values}

\item{id}{What column is not being reorganized and needs to be preserved. Usually "Subject"}

\item{fill}{Passed to spread() fill parameter}

\item{variables_combine.name}{If using more than one variables column then specify name of a new combined column}

\item{x}{dataframe}

\item{variables}{The variable used for spreading}

\item{variables_combine.name}{If using more than one variables column then specify name of a new combined column}

\item{variables_combine.sep}{if using more than one variables column then specifiy how values should be seperated when combined}

\item{values}{A vector of columns that contain the values}

\item{id}{What column is not being reorganized and needs to be preserved. Usually "Subject"}

\item{fill}{Passed to spread() fill parameter}
}
\description{
This function is an extension of the spread() function in tidyverse.
It can multiple values to apply spread() on.

This function is an extension of the spread() function in tidyverse.
It can multiple values to apply spread() on.
}
\examples{
reshape.long(x, variables = "variable", values = c("value1", "value2"), by = "Subject")
reshape.long(x, variables = "variable", values = c("value1", "value2"), by = "Subject")
}
