% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{reshape}
\alias{reshape}
\title{Data Transformation Function}
\usage{
reshape(x, key = "", key_combine.name = "", key_combine.sep = "",
  values = c(), by = "", fill = NA)
}
\arguments{
\item{x}{dataframe}

\item{key}{The variable used for spreading}

\item{key_combine.name}{If using more than one key column then specify name of a new combined column}

\item{key_combine.sep}{if using more than one key column then specifiy how values should be seperated when combined}

\item{values}{A vector of columns that contain the values}

\item{by}{What column is not being reorganized and needs to be preserved. Usually "Subject"}

\item{fill}{Passed to spread() fill parameter}
}
\description{
This function is an extension of the spread() function in tidyverse.
It can multiple values to apply spread() on.
}
\examples{
reshape(x, key = "variable", values = c("value1", "value2"), by = "Subject")
}
\keyword{reshape}
