% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.spread.R
\name{reshape.spread-deprecated}
\alias{reshape.spread-deprecated}
\title{Reshape dataframe from long to wide}
\usage{
reshape.spread(x, variables = "", variables_combine.sep = "_",
  values = c(), id = "", fill = NA)
}
\arguments{
\item{x}{dataframe}

\item{variables}{The variable used for spreading}

\item{variables_combine.name}{If using more than one variables column then specify name of a new combined column}

\item{variables_combine.sep}{if using more than one variables column then specifiy how values should be seperated when combined}

\item{values}{A vector of columns that contain the values}

\item{id}{What column is not being reorganized and needs to be preserved. Usually "Subject"}

\item{fill}{Passed to spread() fill parameter}
}
\description{
This function is an extension of the `tidyr::spread()` function
It can `tidyr::spread()` on multiple keys and values.
}
\seealso{
\code{\link{datawrangling-deprecated}}
}
\keyword{internal}
