% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_gather.R
\name{reshape_gather}
\alias{reshape_gather}
\title{Reshape dataframe from wide to long}
\usage{
reshape_gather(x, variable.names, values, id = NULL,
  separate.pattern = NULL)
}
\arguments{
\item{x}{dataframe}

\item{variable.names}{The variable used for gathering}

\item{values}{If using more than one variables column then specify name of a new combined column}

\item{id}{What column is not being reorganized and needs to be preserved. Usually "Subject"}

\item{separate.pattern}{stuff}
}
\description{
This function is an extension of the `tidyr::gather()` function.
It can `tidyr::gather()` on multiple keys and values
}
\examples{
reshape_gather(x, variables = "variable", values = c("value1", "value2"), by = "Subject")
}
\keyword{reshape}
