% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{A Data Transformation Function}
\usage{
se(x, na.rm = FALSE)
}
\arguments{
\item{x}{vector of values to calculate the se}

\item{na.rm}{Logical. Remove missing values in calculation?}
}
\description{
Calculates the standard error of the mean
}
\examples{
se(x)
}
\keyword{se}
