% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{A Data Transformation Function}
\usage{
trim(x, variables = c(), cutoff = 3.5, context = "", replace = "NA")
}
\arguments{
\item{x}{dataframe}

\item{variables}{variables c() to be trimmed}

\item{cutoff}{zscore cutoff to use for trimming (default: 3.5)}

\item{context}{Name of column to group by}
}
\description{
This function will trim a variable(s) based on a zscore cutoff
}
\examples{
trim(x, variables = c(), cutoff = 3.5)
}
\keyword{trim}
