% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center}
\alias{center}
\title{Create a column of centered variables}
\usage{
center(x, variables = c(colnames(x)), standardize = FALSE,
  drop = FALSE, suffix = NULL)
}
\arguments{
\item{x}{dataframe}

\item{variables}{c() of columns to standardize}

\item{standardize}{Logical. Do you want to calculate
zscores? (Default = FALSE)}

\item{drop}{Drop original non-centered variables}

\item{suffix}{Suffix to add at the end of the column names. Default = NULL}
}
\description{
This function centers a variable around the mean. To create z-scores,
specify `standardize=TRUE`
A new column will be created with the centered values.
There is also the option to center within context.
}
