% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite.R
\name{composite}
\alias{composite}
\title{Create a compositve variable}
\usage{
composite(
  x,
  variables,
  type = "mean",
  standardize = TRUE,
  name = NULL,
  missing.allowed = NULL
)
}
\arguments{
\item{x}{dataframe}

\item{variables}{c() of columns to average together to create the composite}

\item{type}{What type of composite should be calculated?, i.e. mean or sum. (Default = "mean").}

\item{standardize}{Logical. Do you want to calculate the composite based on standardize (z-score) values? (Default = FALSE)}

\item{name}{Name of the new composite variable}

\item{missing.allowed}{Criteria for how many variables can having missing values and still calculate a composite}
}
\description{
This function will create a composite variable
}
\examples{
composite(x, variable = c(), type = "mean", name = "name")
}
\keyword{composite}
