% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates_check.R
\name{duplicates_check}
\alias{duplicates_check}
\title{Check and remove duplicate ids}
\usage{
duplicates_check(
  x,
  id = "Subject",
  unique = c("SessionDate", "SessionTime"),
  save_as = NULL
)
}
\arguments{
\item{x}{dataframe}

\item{id}{Subject ID variable name.}

\item{unique}{Column names that are unique and should be used to
check for duplicate id's}

\item{save_as}{Folder path and file name to output the duplicate ID's}

\item{n}{Number of unique id's expected}

\item{remove}{logical. Remove duplicate ids from data?}
}
\description{
This function checks and removes duplicate ids
}
