% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.latent.R
\name{remove.latent}
\alias{remove.latent}
\title{A Data Cleaning Function}
\usage{
remove.latent(x, factor.list, missing.allowed, id, removed.dir = NULL,
  removed.file = NULL)
}
\arguments{
\item{x}{dataframe}

\item{factor.list}{list of factors and tasks. ex. `list(WMC = c("OSpan", "RotSpan", SymSpan"), Gf = c("RAPM", "NumberSeries", "LetterSets"))`}

\item{missing.allowed}{Proportion of tasks allowed to be missing}

\item{id}{Subject ID variable}

\item{removed.dir}{File directory to save removed subjects to}

\item{removed.file}{File name to save removed subjects to}
}
\description{
This function will remove subjects based on amount of missing data per latent construct
}
\keyword{remove}
