% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.latent.R
\name{remove.latent}
\alias{remove.latent}
\title{A Data Transformation Function}
\usage{
remove.latent(x, factor.list, missing.allowed, id, removed.dir = NULL,
  removed.file = NULL)
}
\arguments{
\item{x}{dataframe}

\item{factor.list}{list of factors and tasks}

\item{missing.allowed}{% of tasks allowed to be missing}

\item{id}{Subject ID variable}

\item{removed.dir}{file directory to save removed subjects to}

\item{removed.file}{file name to save removed subjects to}
}
\description{
This function will remove subjects based on amount of missing data per latent construct
}
\examples{
remove.latent(x, variables = c(), cutoff = 3.5)
}
\keyword{remove}
