% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.save.R
\name{remove.save}
\alias{remove.save}
\title{A Data Transformation Function}
\usage{
remove.save(x, remove, output.dir = NULL, output.file = NULL)
}
\arguments{
\item{x}{dataframe}

\item{remove}{the dataframe that contains subjects to be removed}

\item{output}{folder directory path to save removed data}
}
\description{
This function centers a variable around the mean. A new column will be created with the
centered values. There is also the option to center within context.
}
\examples{
remove.save(data, remove = data_remove, save = "data/remove", taskname = "Flanker")
}
\keyword{remove}
