% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.gather.R, R/reshape.long.R
\name{reshape.gather}
\alias{reshape.gather}
\title{Data Transformation Function}
\usage{
reshape.gather(x, variable.name, values, id = NULL,
  separate.into = NULL, separate.pattern = NULL)

reshape.gather(x, variable.name, values, id = NULL,
  separate.into = NULL, separate.pattern = NULL)
}
\arguments{
\item{x}{dataframe}

\item{variable.name}{The variable used for spreading}

\item{values}{If using more than one variables column then specify name of a new combined column}

\item{id}{What column is not being reorganized and needs to be preserved. Usually "Subject"}

\item{separate.into}{Passed to spread() fill parameter}

\item{separate.pattern}{stuff}

\item{variable.names}{The variable used for spreading}

\item{x}{dataframe}

\item{values}{If using more than one variables column then specify name of a new combined column}

\item{id}{What column is not being reorganized and needs to be preserved. Usually "Subject"}

\item{separate.pattern}{stuff}
}
\description{
This function is an extension of the spread() function in tidyverse.
It can multiple values to apply spread() on.

This function is an extension of the spread() function in tidyverse.
It can multiple values to apply spread() on.
}
\examples{
reshape.gather(x, variables = "variable", values = c("value1", "value2"), by = "Subject")
reshape.gather(x, variables = "variable", values = c("value1", "value2"), by = "Subject")
}
\keyword{reshape}
